wip heuristic, almost finish, need a patch on capture coef

+++ GomokuLib/GomokuLib/Typing/__init__.py
@@ -8,7 +8,7 @@ BoardDtype = np.int8
 ActionDtype = np.int8
 GameZoneDtype = np.int8
 PruningDtype = np.int8
-HeuristicGraphDtype = np.int8
+HeuristicGraphDtype = np.float32
 
 TupleDtype = np.int32
 MCTSIntDtype = np.int32
